Move InvalidRSAPrivateKeyException to `service` directory; fix exception declaration in TokenAuthentication methods
